requirements.txt
Django==2.0.7
django-guardian==1.4.9
django-widget-tweaks==1.4.2
Pillow==5.2.0
pytz==2018.5
six==1.11.0
